Updated Drawers
updated drawers and cleaned redundant scripts
