--- Team MMA Railshooter NavTool/Assets/Editor/ScriptNavigationEditor.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 2ad390eb34a65744db9a2d1b9a6861ee
-timeCreated: 1442073981
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
