--- Team MMA Railshooter NavTool/Assets/Editor/WaypointEditorDrawer.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 07d65b18c69345e46b14df35ac974942
-timeCreated: 1441901493
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
