+++ Team MMA Railshooter NavTool/Assets/Scripts/ScriptModSupport.cs
@@ -11,7 +11,7 @@ public class ScriptModSupport : MonoBehaviour {
     public GameObject movementWaypoint;
     public GameObject facingWaypoint;
     public GameObject effectWaypoint;
-    public ScriptNavigation player;
+    public ScriptEngine player;
 
 
     string defaultModFileText = "Edit this file for modding timelines and delete this line";
@@ -23,7 +23,7 @@ public class ScriptModSupport : MonoBehaviour {
     {
         if (player == null)
         {
-            player = GameObject.FindGameObjectWithTag("Player").GetComponent<ScriptNavigation>();
+            player = GameObject.FindGameObjectWithTag("Player").GetComponent<ScriptEngine>();
             if (player == null)
             {
                 ScriptErrorLogging.logError("No Player Object found, please add a player to the scene and check the tag.");
