--- Team MMA Railshooter NavTool/Assets/Scripts/ScriptNavigation.cs
@@ -1,22 +1,0 @@
-﻿using UnityEngine;
-using System.Collections;
-
-/// <summary>
-/// @Author Marshall Mason
-/// </summary>
-public class ScriptNavigation : MonoBehaviour {
-
-
-    public ScriptWaypoints[] waypoints;
-
-	// Use this for initialization
-	void Start () {
-	
-	}
-	
-	// Update is called once per frame
-	void Update ()
-    {
-	
-	}
-}
