--- Team MMA Railshooter NavTool/Assets/Scripts/ScriptNavigation.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: d355873ff424e5049808200d2615d784
-timeCreated: 1442073801
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
